severity の仕訳処理を共通化
